Crate registry_pol

The Group Policy Object Editor stores registry-based configuration settings in two Registry.pol files, stored in folders under the <drive>:\Windows\System32\GroupPolicy\ folder. One file contains computer settings and the other file contains user settings. The Group Policy Object Editor saves the settings to these files on exit, and imports the settings on startup.

This library does not and will never depend on any Windows-specific libraries and will therefore work on any system.

Reexports

pub use self::v1::*;

Modules

v1
Registry Policy File format version 1.

Constants

REGFILE_SIGNATURE
First part of header, equivalent to "PReg" in little-endian.