Struct registry_pol::v1::RegistryValue

pub struct RegistryValue {
    pub key: String,
    pub value: Option<String>,
    pub data_type: Option<RegistryValueType>,
    pub data: Option<Vec<u8>>,
}

A representation of registry value version 1.

If value, type, size, or data are missing or zero, only the registry key is created.

Fields
key: String

Path to the registry key.

Do not include HKEY_LOCAL_MACHINE or HKEY_CURRENT_USER in the registry path. The location of the file determines which of these keys are used.

value: Option<String>

The name of the registry value.

The following values have special meaning for this field:

Value 	Meaning

**DeleteValues 	A semicolon-delimited list of values to delete. Use as a value of the associated key.

**Del.<valuename> 	Deletes a single value. Use as a value of the associated key.

**DelVals 	Deletes all values in a key. Use as a value of the associated key.

**DeleteKeys 	A semicolon-delimited list of keys to delete.

**SecureKey 	See Access Rights and Access Masks.

data_type: Option<RegistryValueType>

The data type.

data: Option<Vec<u8>>

The user-supplied data.